requirements.txt
streamlit
pyvis
langchain-ollama
langchain-huggingface
langchain-community
langchain-text-splitters
python-dotenv
neo4j
sentence-transformers
networkx
numpy
scipy
torch
biopython
